+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 4.4.0 &mdash; 2021-10-25
+* [Added] New `Authorization.strategy "SAS"` (alternative to default `JWT`).
+* [Fixed] Fixed __`RCTEventEmitter`__ warning "addListener", "removeListeners".  A corresponding fix has also been released for [react-native-background-fetch](https://github.com/transistorsoft/react-native-background-fetch/blob/master/CHANGELOG.md#404--2021-10-22) so be sure to update to `background-fetch >= 4.0.4`.
+* [Changed] **Deprecated** `BackgroundGeolocation.removeListener`.  All event-handlers now return a `react-native` `Subscription` instance containing a `.remove()` method.  You will keep track of your own `subscription` instances and call `.remove()` upon them when you wish to remove an event listener.  Eg:
+
+```javascript
+/// OLD
+const onLocation = (location) => {
+    console.log('[onLocation');
+}
+BackgroundGeolocation.onLocation(onLocation);
+...
+// deprecated: removeListener
+BackgroundGeolocation.removeListener('location', onLocation);
+
+/// NEW:  capture returned subscription instance.
+const onLocationSubscription = BackgroundGeolocation.onLocation(onLocation);
+...
+// Removing an event-listener.
+onLocationSubscription.remove();
+```
+
 ## 4.3.0 &mdash; 2021-09-14
 
 * [Added][Android] Implement new Android 12 "reduced accuracy" mechanism `requestTemporaryFullAccuracy`.